test(harness): move TestDatabricksCompletion to chat_live_longtail
Audit 5 marks the class MIGRATE (live BaseLLMChatTest + BaseAnthropicChatTest
subclass) though the 8c suite mapping omits databricks; chat_live_longtail is
the env-gated long-tail suite, so it lands there. Only the offline
api-base/api-key error-path test remains in test_databricks.py.
