+++ tests/llm_translation/test_databricks.py
@@ -8,7 +8,6 @@ sys.path.insert(
 )  # Adds the parent directory to the system path
 import litellm
 from litellm.exceptions import BadRequestError
-from base_llm_unit_tests import BaseLLMChatTest, BaseAnthropicChatTest
 
 
 @pytest.mark.parametrize("set_base", [True, False])
@@ -48,18 +47,4 @@ def test_throws_if_api_base_or_api_key_not_set_without_databricks_sdk(
         )
     assert any(msg in str(exc) for msg in err_msg)
 
-
-@pytest.mark.skip(reason="Databricks rate limit errors")
-class TestDatabricksCompletion(BaseLLMChatTest, BaseAnthropicChatTest):
-    def get_base_completion_call_args(self) -> dict:
-        return {"model": "databricks/databricks-claude-3-7-sonnet"}
-
-    def get_base_completion_call_args_with_thinking(self) -> dict:
-        return {
-            "model": "databricks/databricks-claude-3-7-sonnet",
-            "thinking": {"type": "enabled", "budget_tokens": 1024},
-        }
-
-    def test_pdf_handling(self, pdf_messages):
-        pytest.skip("Databricks does not support PDF handling")
